changelog 1.1.31

+++ app/api/cargo/delete/route.ts
@@ -1,6 +1,13 @@
-import prisma from '@/lib/prisma';
+import { validateRequest } from "@/auth";
+import prisma from "@/lib/prisma";
 
 export async function DELETE(req: Request) {
+  const { user } = await validateRequest();
+
+  if (!user) {
+    return new Response("Unauthorized", { status: 401 });
+  }
+
   const { id } = await req.json();
   console.log(id);
   const post = await prisma.cargo.delete({
