+++ app/api/escala/delete/mercadoria/route.ts
@@ -1,6 +1,13 @@
-import prisma from '@/lib/prisma';
+import { validateRequest } from "@/auth";
+import prisma from "@/lib/prisma";
 
 export async function DELETE(req: Request) {
+  const { user } = await validateRequest();
+
+  if (!user) {
+    return new Response("Unauthorized", { status: 401 });
+  }
+
   const { id } = await req.json();
   const result = await prisma.relac_mercadoria_escala.delete({
     where: {
@@ -9,7 +16,7 @@ export async function DELETE(req: Request) {
   });
 
   if (!result) {
-    return new Response('Mercadoria não encontrada', { status: 404 });
+    return new Response("Mercadoria não encontrada", { status: 404 });
   }
 
   return Response.json(result);
